requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Flask==3.0.3
numpy==2.0.0
openpyxl==3.1.3
qq-botpy==1.2.1
requests==2.32.3
scipy==1.13.1
tqdm==4.66.4
